requirements_locals.txt
autopep8==1.5.7
certifi==2020.12.5
chardet==4.0.0
Django==2.2.3
django-autoslug==1.9.8
idna==2.10
pycodestyle==2.7.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
